+++ backend/constants.mjs
@@ -1,6 +1,8 @@
-export const db_config = {
-    host: process.env.DB_HOST,
-    user: process.env.DB_USER,
-    password: process.env.DB_PASSWORD,
-    database: process.env.DB_NAME,
-  };
+import { fileURLToPath } from 'url';
+import { dirname, join } from 'path';
+
+const root_folder_path = dirname(dirname(fileURLToPath(import.meta.url)));
+
+export const db_path = join(root_folder_path,'data','gamebox.db');
+
+export const static_folder_path = join(root_folder_path,'static');