+++ backend/models/models.mjs
@@ -1,97 +1,110 @@
-import mysql from 'mysql2/promise';
-import { db_config } from "../constants.mjs";
+import { connectDB } from '../controllers/databaseController.mjs'
+export async function createTableRow(table_name,columns_names,columns_values){
+    let db = connectDB();
+    
+    let query = db.prepare(`INSERT INTO ${table_name} (${columns_names.join(',')}) VALUES (${columns_values.map(() => '?').join(',')})`);
+    
+    return new Promise((resolve,reject) => {
+        query.run(columns_values, function (err) {
+            query.finalize();
+            db.close();
+            if (err) {
+                console.error('Error while running createTableRow query',err);
+                reject(err);
+            }
+            else{
+                resolve(true);
+            }
+        });
+    });
+}
 
 
-export async function createTableRow(table,columns_names,columns_values){
-    let success = true;
-    let query_txt = `INSERT INTO ${table_name} (${columns_names.join(',')}) VALUES (${Array.from({ length: n }, () => '?').join(',')})`;
-    try {
-        const connection = mysql.createConnection(db_config);
-        await connection.execute(query_txt,columns_values);
-        await connection.end();
-    }
-    catch (error) {
-        success = false;
-        console.error(`Could not create new row in table ${table}`);
-    }
+export async function getTableRowIdMatchingColumnValue(table_name,column_name,column_value){
+    let db = connectDB();
 
-    return success;
+    return new Promise((resolve,reject) => {
+        db.get(`SELECT "id" FROM ${table_name} WHERE ${column_name} = ?`,[row_id],(err,row) =>{
+            db.close();
+            if (err){
+                reject(err);
+            }
+            else{
+                resolve(row ? row[column_name] : null);
+            }
+        });
+    });
 }
 
 
 export async function getTableRowColumn(table_name,column_name,row_id) {
-    let column_value = null;
-    try {
-        const connection = await mysql.createConnection(db_config);
-        const [rows] = await connection.execute(`SELECT ${column_name} FROM ${table_name} WHERE "id" = ?`, [row_id]);
-        await connection.end();
-
-        if (rows.length > 0) {
-            column_value = rows[0][column_name];
-        }
-    }
-    catch (error) {
-        console.error(`Error while retrieving ${table_name}, id=${row_id}, column ${column_name}: `, error);
-    }
-
-    return column_value;
+    let db = connectDB();
+    
+    return new Promise((resolve, reject) => {
+        db.get(`SELECT ${column_name} FROM ${table_name} WHERE "id" = ?`,[row_id],(err,row) =>{
+            db.close();
+            if (err){
+                reject(err);
+            }
+            else{
+                resolve(row ? row[column_name] : null);
+            }
+        });
+    });
 }
 
 
 export async function getTableRowColumns(table_name,columns_names,row_id){
-    let columns_values = null;
-    try {
-        const connection = await mysql.createConnection(db_config);
-        const [rows] = await connection.execute(`SELECT ${columns_names.join(',')} FROM ${table_name} WHERE "id" = ?`, [row_id]);
-        await connection.end();
+    let db = connectDB();
 
-        if (rows.length > 0) {
-            columns_values = [];
-            for (let column_name of columns_names){
-                columns_values.push(rows[0][column_name]);
+    return new Promise((resolve,reject) => {
+        db.get(`SELECT ${columns_names.join(',')} FROM ${table_name} WHERE "id" = ?`, [row_id], (err,rows) => {
+            db.close();
+            if (err){
+                reject(err);
             }
-        }
-    }
-    catch (error) {
-        console.error(`Error while retrieving ${table_name}, id=${row_id}, columns ${columns_names.join(',')}: `, error);
-    }
-
-    return columns_values;
+            else{
+                resolve(row ? columns_names.map(column_name => row[column_name]) : null);
+            }
+        });
+    });
 }
 
 
 export async function deleteTableRowMatchingColumns(table_name,columns_names,columns_values) {
-    let success = true;
-    try {
-        let intermediate_text = '';
-        for (let column_name of columns_names){
-            intermediate_text += `${column_name} = ? AND `;
-        }
-        intermediate_text = intermediate_text.substring(0,intermediate_text-5);
+    let db = connectDB();
 
-        const connection = await mysql.createConnection(db_config);
-        await connection.execute(`DELETE FROM ${table_name} WHERE ${intermediate_text}`, columns_values);
-        await connection.end();
-    }
-    catch (error){
-        console.log(`Error while deleting row from table ${table_name}`);
-        success = false;
-    }
-
-    return success;
+    let query = db.prepare(`DELETE FROM ${table_name} WHERE ${columns_names.map(column_name => `${column_name} = ?`).join(' AND ')}`)
+    return new Promise((resolve,reject) => {
+        query.run(columns_values, function(err){
+            query.finalize();
+            db.close();
+            if (err){
+                reject(err);
+            }
+            else{
+                resolve(true);
+            }
+        });
+    });
 }
 
 
-export async function editTableRowColumn(table_name,row_id,column_name,column_value){
-    let success = true;
-    try{
-        const connection = await mysql.createConnection(db_config);
-        await connection.execute(`UPDATE ${table_name} SET ${column_name} = ? WHERE "id" = ?`,[column_name,column_value]);
-        await connection.end();
-    }
-    catch (error){
-        console.error(`Could not edit ${table_name} ${row_id} ${column_name}`);
-        success = false;
-    }
-    return success;
+export function editTableRowColumn(table_name,row_id,column_name,column_value){
+    let db = connectDB();
+    let query = db.prepare(`UPDATE ${table_name} SET ${column_name} = ? WHERE "id" = ?`);
+
+    return new Promise((resolve,reject) => {
+        query.run([column_value,row_id], function(err){
+            query.finalize();
+            db.close();
+            if (err){
+                console.error(`Could not edit ${table_name} ${row_id} ${column_name}`, err);
+                reject(err);
+            }
+            else{
+                resolve(true);
+            }
+        });
+    });
 }