+++ backend/models/scores.mjs
@@ -1,6 +1,5 @@
-import mysql from 'mysql2/promise';
+import { connectDB } from '../controllers/databaseController.mjs'
 import { createTableRow, deleteTableRowMatchingColumns, editTableRowColumn } from "./models.mjs";
-import { db_config } from '../constants.mjs';
 
 async function createScore(value,game_id,user_id) {
     return await createTableRow('Scores',['"value"','"game_id"','"user_id"'],[value,game_id,user_id]);
@@ -23,17 +22,17 @@ async function editScore(score_id,value) {
 
 
 async function getGameScores(game_id,max_index){
-    let success = true;
-    let query = `SELECT *FROM Scores WHERE "game_id" = ? ORDER BY "value" DESC LIMIT ${max_index};`
-    try {
-        const connection = await mysql.createConnection(db_config);
-        await connection.execute(query,[game_id]);
-        await connection.end();
-    }
-    catch (error) {
-        success = false;
-        console.error(`Could not retrieve scores for Game: ${game_id}`,error);
-    }
-
-    return success;
+    let db = connectDB();
+    
+    return new Promise((resolve,reject) => {
+        db.all(`SELECT "value","user_id" FROM Scores WHERE "game_id" = ? ORDER BY "value" DESC LIMIT ?;`,[game_id,max_index],(err,rows) => {
+            db.close();
+            if (err) {
+                reject(err);
+            }
+            else{
+                resolve(rows);
+            }
+        });
+    });
 }