+++ index.mjs
@@ -1,19 +1,24 @@
+import dotenv from 'dotenv';
+dotenv.config();
 import express from 'express'
-import { fileURLToPath } from 'url';
-import { dirname, join } from 'path';
-import game_router from './backend/routes/games.mjs';
+import games_router from './backend/routes/games.mjs';
+import { static_folder_path } from './backend/constants.mjs';
+import { initDatabase } from './backend/controllers/databaseController.mjs';
+import auth_router from './backend/routes/auth.mjs';
 
 const app = express();
 // ### ROUTES ###
 
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = dirname(__filename);
-app.use("/static", express.static(join(__dirname, '/static')));
+app.use("/static", express.static(static_folder_path));
 app.get('/', (req, res) => {
-    res.redirect(301, '/static/frontend/pong/index.html');
+    res.redirect(301, '/static/login/index.html');
 });
 
-app.use('/games',game_router);
+app.use('/games',games_router);
+
+app.use('/auth',auth_router);
+
+initDatabase();
 
 // ### HOSTING ###
 const hostname = 'localhost';