+++ static/400m/js/constants.mjs
@@ -1,0 +1,14 @@
+/* JavaScript FRONT
+ * Project: 400m
+ * Package: js
+ * Module: constants
+ * Usage: General 400m constants.
+ * Date: 15/11/2023
+ * author: BoxBoxJason
+ */
+export const maxRound = 4;
+export const numberDicesPerThrow = 2;
+export const maxThrows = 9;
+export const minDiceValue = 1;
+export const maxDiceValue = 6;
+export const maxDistance = 400;