+++ static/frontend/pong/js/constants.mjs
@@ -1,7 +1,0 @@
-// Game Frame
-export const game_frame = document.getElementById('gameFrame');
-// Paddles
-export const paddle1 = document.getElementById("paddle1");
-export const paddle2 = document.getElementById("paddle2");
-// Ball
-export const ball = document.getElementById("ball");