+++ static/pong/js/constants.mjs
@@ -1,0 +1,15 @@
+/* JavaScript FRONT
+ * Project: Pong
+ * Package: js
+ * Module: constants
+ * Usage: Pong related constants (front end div).
+ * Date: 15/10/2023
+ * author: BoxBoxJason
+ */
+// Game Frame
+export const game_frame = document.getElementById('gameFrame');
+// Paddles
+export const paddle1 = document.getElementById("paddle1");
+export const paddle2 = document.getElementById("paddle2");
+// Ball
+export const ball = document.getElementById("ball");