+++ CHANGELOG.md
@@ -1,6 +1,32 @@
 # GameBox CHANGELOG
 
-## GameBox V1.0.0 Pong release 01/12/2023
+## GameBox V1.0.2 OLYMPIC 400M RELEASE (?/12/2023)
+
+### New Features
+- Released Olympic 400m minigame
+
+### Fixes
+-
+
+### TODO
+1. Implement sessions to keep users 
+2. Implement new games
+
+## GameBox V1.0.1 AUTHENTIFICATION RELEASE (04/12/2023)
+
+### New features
+- Released authentification page and linked users table
+
+### Fixes
+- Changed static files tree
+- Changed routes to increase clarity
+
+### TODO
+1. Implement home page
+2. Implement sessions to keep users logged in
+3. Implement new games
+
+## GameBox V1.0.0 PONG RELEASE (01/12/2023)
 
 ### New features
 - Released fully functionnal Pong Game, (details explained in [README.md](./README.md "README"))